refactor(sync): Split deletion planning helpers
SyncTrunk coordinates trunk synchronization, merged-branch deletion, and
upstack retargeting in one workflow. Move the deletion planning helpers into
deletion.go and the forge retarget helpers into retarget.go so each helper
group sits with its local invariants while handler.go keeps the workflow entry
point.

This is a package-local file organization change. Branch deletion behavior,
restack mode handling, and forge retargeting behavior remain unchanged.
